Refactor methods to return algorithm properties
Modified "getAllAlgorithmInfo()" in Controllers.java to return a map of algorithm properties instead of BaseSort objects. This allows for a more detailed and useful representation of the different sorting algorithms utilized in the code. Also adjusted the "readPropsFile()" method in PropsUtils.java to include the algorithm name as part of the returned properties.