Refactor NearlySortedArray with sorting and random swaps
The NearlySortedArray class has been refactored for generating and sorting an array of random integers. The generated array is then subjected to a random number of swaps for producing a 'nearly sorted' array as intended. The swapping operation has been optimized to reduce redundancy, improving the overall efficiency of the method.