Add ShellSort algorithm and improve variable highlighting
The ShellSort algorithm is implemented and added to the existing sorting algorithms. Additionally, improvements have been made to how variables are highlighted and displayed in the sorting algorithms. The concept of "highlightVariable" has replaced the previous "highlightValue" and a new "displayVariable" method has been introduced as part of these changes.