Refactor BubbleSort and BubbleSortOptimised classes
The private method 'optimized' in BubbleSortOptimised class and 'standart' in BaseSort class were deleted and their contents were moved directly into the override 'run' methods. This was done to eliminate unnecessary method calls and streamline the code. The method 'populateProps' was also re-positioned for better readability and consistency across classes.